Split MemoryFailPoint into Unix and Windows versions.
