+++ src/System.Private.CoreLib/shared/System/Runtime/MemoryFailPoint.cs
@@ -13,14 +13,10 @@
 ** 
 ===========================================================*/
 
-using System;
 using System.IO;
-using Microsoft.Win32;
-using System.Runtime.InteropServices;
 using System.Threading;
 using System.Runtime.CompilerServices;
 using System.Runtime.ConstrainedExecution;
-using System.Runtime.Versioning;
 using System.Diagnostics;
 
 /* 
@@ -156,7 +152,6 @@ namespace System.Runtime
             if (sizeInMegabytes <= 0)
                 throw new ArgumentOutOfRangeException(nameof(sizeInMegabytes), SR.ArgumentOutOfRange_NeedNonNegNum);
 
-#if !FEATURE_PAL // Remove this when CheckForAvailableMemory is able to provide legitimate estimates
             ulong size = ((ulong)sizeInMegabytes) << 20;
             _reservedMemory = size;
 
@@ -191,7 +186,11 @@ namespace System.Runtime
             // would probably work, but do some thinking first.)
             for (int stage = 0; stage < 3; stage++)
             {
-                CheckForAvailableMemory(out availPageFile, out totalAddressSpaceFree);
+                if (!CheckForAvailableMemory(out availPageFile, out totalAddressSpaceFree))
+                {
+                    // _mustSubtractReservation == false
+                    return;
+                }
 
                 // If we have enough room, then skip some stages.
                 // Note that multiple threads can still lead to a race condition for our free chunk
@@ -312,80 +311,7 @@ namespace System.Runtime
 
             AddMemoryFailPointReservation((long)size);
             _mustSubtractReservation = true;
-#endif
         }
-
-#if !FEATURE_PAL
-        private static void CheckForAvailableMemory(out ulong availPageFile, out ulong totalAddressSpaceFree)
-        {
-            bool r;
-            Interop.Kernel32.MEMORYSTATUSEX memory = new Interop.Kernel32.MEMORYSTATUSEX();
-            r = Interop.Kernel32.GlobalMemoryStatusEx(ref memory);
-            if (!r)
-                throw Win32Marshal.GetExceptionForLastWin32Error();
-            availPageFile = memory.availPageFile;
-            totalAddressSpaceFree = memory.availVirtual;
-            // Console.WriteLine($"Memory gate:  Mem load: {memory.memoryLoad}%  Available memory (physical + page file): {(memory.availPageFile >> 20)} MB  Total free address space: {memory.availVirtual >> 20} MB  GC Heap: {(GC.GetTotalMemory(true) >> 20)} MB");
-        }
-
-        // Based on the shouldThrow parameter, this will throw an exception, or 
-        // returns whether there is enough space.  In all cases, we update
-        // our last known free address space, hopefully avoiding needing to 
-        // probe again.
-        private static unsafe bool CheckForFreeAddressSpace(ulong size, bool shouldThrow)
-        {
-            // Start walking the address space at 0.  VirtualAlloc may wrap
-            // around the address space.  We don't need to find the exact
-            // pages that VirtualAlloc would return - we just need to
-            // know whether VirtualAlloc could succeed.
-            ulong freeSpaceAfterGCHeap = MemFreeAfterAddress(null, size);
-
-            // Console.WriteLine($"MemoryFailPoint: Checked for free VA space.  Found enough? {(freeSpaceAfterGCHeap >= size)}  Asked for: {size}  Found: {freeSpaceAfterGCHeap}");
-
-            // We may set these without taking a lock - I don't believe
-            // this will hurt, as long as we never increment this number in 
-            // the Dispose method.  If we do an extra bit of checking every
-            // once in a while, but we avoid taking a lock, we may win.
-            LastKnownFreeAddressSpace = (long)freeSpaceAfterGCHeap;
-            LastTimeCheckingAddressSpace = Environment.TickCount;
-
-            if (freeSpaceAfterGCHeap < size && shouldThrow)
-                throw new InsufficientMemoryException(SR.InsufficientMemory_MemFailPoint_VAFrag);
-            return freeSpaceAfterGCHeap >= size;
-        }
-
-        // Returns the amount of consecutive free memory available in a block
-        // of pages.  If we didn't have enough address space, we still return 
-        // a positive value < size, to help potentially avoid the overhead of 
-        // this check if we use a MemoryFailPoint with a smaller size next.
-        private static unsafe ulong MemFreeAfterAddress(void* address, ulong size)
-        {
-            if (size >= s_topOfMemory)
-                return 0;
-
-            ulong largestFreeRegion = 0;
-            Interop.Kernel32.MEMORY_BASIC_INFORMATION memInfo = new Interop.Kernel32.MEMORY_BASIC_INFORMATION();
-            UIntPtr sizeOfMemInfo = (UIntPtr)Marshal.SizeOf(memInfo);
-
-            while (((ulong)address) + size < s_topOfMemory)
-            {
-                UIntPtr r = Interop.Kernel32.VirtualQuery(address, ref memInfo, sizeOfMemInfo);
-                if (r == UIntPtr.Zero)
-                    throw Win32Marshal.GetExceptionForLastWin32Error();
-
-                ulong regionSize = memInfo.RegionSize.ToUInt64();
-                if (memInfo.State == Interop.Kernel32.MEM_FREE)
-                {
-                    if (regionSize >= size)
-                        return regionSize;
-                    else
-                        largestFreeRegion = Math.Max(largestFreeRegion, regionSize);
-                }
-                address = (void*)((ulong)address + regionSize);
-            }
-            return largestFreeRegion;
-        }
-#endif
 
         ~MemoryFailPoint()
         {
